Suppressing compilation warnings from utilities: * Explicit conversion to integer. * Unused variable. * Bugfix: complex*4 was used instead of complex*8. * Using our standard constant for 4.

+++ src/utils/casida_spectrum.F90
@@ -185,7 +185,7 @@ contains
     character(len=256) :: string
     logical :: is_complex
 
-    nsteps = (cs%max_energy - cs%min_energy) / cs%energy_step
+    nsteps = int((cs%max_energy - cs%min_energy) / cs%energy_step)
     SAFE_ALLOCATE(spectrum(1:cs%space%dim+1, 1:nsteps))
     spectrum = M_ZERO
